Updated to 26.1 and used BigInteger for economy. Increased balance cap from 2^31 to 2^63.

+++ src/main/java/rickiewars/guishop/api/database/DatabaseManager.java
@@ -7,11 +7,11 @@ public interface DatabaseManager {
     void updateAccount(String currency, String uuid, String name);
     String getName(String currency, String uuid);
 
-    int getBalance(String currency, String uuid);
-    boolean setBalance(String currency, String uuid, int money);
-    void setAllBalance(String currency, int money);
-    boolean changeBalance(String currency, String uuid, int money);
-    void changeAllBalance(String currency, int money);
+    long getBalance(String currency, String uuid);
+    boolean setBalance(String currency, String uuid, long money);
+    void setAllBalance(String currency, long money);
+    boolean changeBalance(String currency, String uuid, long money);
+    void changeAllBalance(String currency, long money);
 
     List<Map<String, Object>> top(String currency, int topAmount);
     Map<String, String> rank(String currency, int rank);
